# swagger/combinedServerManagerService.yaml
openapi: 3.0.0
info:
  title: Server Manager Service
  description: Define and control your servers with these routes.
  version: 0.1.0
servers:
  - url: /server-manager-service
  - url: /
security:
  - ApiBearerAuth: []
...
paths:
...
  /hosts/{hostId}/servers:
    post:
      tags:
        - servers
      summary: Create a new Server
      operationId: createServer
      parameters:
        - $ref: '#/components/parameters/hostIdInPath'
      requestBody:
        $ref: '#/components/requestBodies/serverCreateRequest'
      responses:
        '200':
          $ref: '#/components/responses/serverResponse'
        '401':
          $ref: '#/components/responses/notAuthorizedError'
        '500':
          $ref: '#/components/responses/internalServerError'
    get:
      tags:
        - servers
      summary: Get all Servers
      operationId: getServers
      parameters:
        - $ref: '#/components/parameters/hostIdInPath'
        - in: query
          name: isUpdatable
          schema:
            type: boolean
          description: filter results based on if a Server is updatable or not
      responses:
        '200':
          $ref: '#/components/responses/serverArrayResponse'
        '401':
          $ref: '#/components/responses/notAuthorizedError'
        '404':
          $ref: '#/components/responses/serverNotFoundError'
        '500':
          $ref: '#/components/responses/internalServerError'
  /hosts/{hostId}/servers/{serverId}:
    get:
      tags:
        - servers
      summary: Get a Server by id
      operationId: getServer
      parameters:
        - $ref: '#/components/parameters/hostIdInPath'
        - $ref: '#/components/parameters/serverIdInPath'
      responses:
        '200':
          $ref: '#/components/responses/serverResponse'
        '401':
          $ref: '#/components/responses/notAuthorizedError'
        '404':
          $ref: '#/components/responses/serverNotFoundError'
        '500':
          $ref: '#/components/responses/internalServerError'
    patch:
      tags:
        - servers
      summary: Patch a Server by id
      operationId: patchServer
      parameters:
        - $ref: '#/components/parameters/hostIdInPath'
        - $ref: '#/components/parameters/serverIdInPath'
      requestBody:
        $ref: '#/components/requestBodies/serverPatchRequest'
      responses:
        '200':
          $ref: '#/components/responses/serverResponse'
        '401':
          $ref: '#/components/responses/notAuthorizedError'
        '404':
          $ref: '#/components/responses/serverNotFoundError'
        '500':
          $ref: '#/components/responses/internalServerError'
    delete:
      tags:
        - servers
      summary: Delete a Server by id
      operationId: deleteServer
      parameters:
        - $ref: '#/components/parameters/hostIdInPath'
        - $ref: '#/components/parameters/serverIdInPath'
      responses:
        '200':
          $ref: '#/components/responses/serverResponse'
        '401':
          $ref: '#/components/responses/notAuthorizedError'
        '404':
          $ref: '#/components/responses/serverNotFoundError'
        '500':
          $ref: '#/components/responses/internalServerError'
...
  /hosts/{hostId}/servers/{serverId}/stop:
    post:
      tags:
        - servers
      summary: Stop a Server by id
      operationId: stopServer
      parameters:
        - $ref: '#/components/parameters/hostIdInPath'
        - $ref: '#/components/parameters/serverIdInPath'
      responses:
        '200':
          $ref: '#/components/responses/serverArrayResponse'
        '401':
          $ref: '#/components/responses/notAuthorizedError'
        '404':
          $ref: '#/components/responses/serverNotFoundError'
        '500':
          $ref: '#/components/responses/internalServerError'
  /hosts/{hostId}/servers/{serverId}/update:
    post:
      tags:
        - servers
      summary: Update a Server by id
      description: This is for rebuilding containers to bring them fully up to date. If you are trying to patch the configuration of a Server, use the PATCH /servers/{serverId} route instead.
      operationId: updateServer
      parameters:
        - $ref: '#/components/parameters/hostIdInPath'
        - $ref: '#/components/parameters/serverIdInPath'
      responses:
        '200':
          $ref: '#/components/responses/serverResponse'
        '401':
          $ref: '#/components/responses/notAuthorizedError'
        '404':
          $ref: '#/components/responses/serverNotFoundError'
        '500':
          $ref: '#/components/responses/internalServerError'
...
components:
  securitySchemes:
    ApiBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    notAuthorizedErrorSchema:
      type: object
      properties:
        description:
          type: string
          example: UnauthorizedError
        message:
          type: string
          example: Requester is not authorized.
...
    internalServerErrorSchema:
      type: object
      properties:
        description:
          type: string
          example: InternalServerError
        message:
          type: string
          example: Your request cannot be processed at this time. The server may be having a bad day.
...
    serverResponseSchema:
      required:
        - id
        - applicationName
        - containerName
        - isUpdatable
        - isInResponseChain
      type: object
      properties:
        id:
          type: string
          format: uuid
        applicationName:
          type: string
        containerName:
          type: string
        isUpdatable:
          type: boolean
        isInResponseChain:
          type: boolean
        hostId:
          type: string
          format: uuid
    serverNotFoundErrorSchema:
      type: object
      properties:
        description:
          type: string
          example: NotFoundError
        message:
          type: string
          example: The server with id 0192de27-1917-702e-93ed-cc4e77227d99 was not found.
    serverRequestSchema:
      required:
        - applicationName
        - containerName
        - hostId
      type: object
      properties:
        applicationName:
          type: string
        containerName:
          type: string
        isUpdatable:
          type: boolean
        isInResponseChain:
          type: boolean
        hostId:
          type: string
...
  responses:
...
    notAuthorizedError:
      description: Unauthorized Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/notAuthorizedErrorSchema'
...
    internalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/internalServerErrorSchema'
...
    serverArrayResponse:
      description: Success
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/serverResponseSchema'
    serverNotFoundError:
      description: The Server could not be found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/serverNotFoundErrorSchema'
    serverResponse:
      description: Success
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/serverResponseSchema'
...
  requestBodies:
...
    serverCreateRequest:
      description: Create a new Server
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/serverRequestSchema'
    serverPatchRequest:
      description: Update an existing Server
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/serverRequestSchema'
...
  parameters:
    hostIdInPath:
      description: id of the Host
      name: hostId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    serverIdInPath:
      description: id of the Server
      name: serverId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...